Merged PR 270: Bugfix in Purging.
Bugfix in Purging.
When there was a purge comment to all the Fluid boards, the _send_purge_par() was not called for every Fluidsystem before it was going to the next step. Like this it purged for ever, as the time was still set to 0, which is used for endless purging.