Merged PR 422: TX-Robot optimized
fix: Implemented a signal to read, if the scanner is powered up, which is getting checked, before sending Scanner-commands for the Robot to the PLC. -> Should solve Cap-2 Problem, but works only, if it's a PLC version, which has this also implemented
fix: Let the waste pump run for the same time as the Flush cycle is running
feat: Implemented a robot_done signal, to let the scanner know, when he needs to release the offset for moving to the robot.
fix: Found mistake in the putty m-Command of the LB702. Now it should be possible to move a motor alone, which was not possible before.
fix: Released the ref_done signal, in case "Reference" is clicked, like this the scanner_enable signal is removed, until the reference is set again.
feat: Removed some useless putty-Commands and information in the Tx-Rob