Merged PR 441: Made TX404 working as TX802
feat: Made the TX404 also working as a TX802 -> Some changes on the Robot were needed for this.
feat: Implemented in TX-Purging and Flushing to check if the scanner is ready befor giving a command
fix: Changed Front and Back motor of the Wrinkle detection on the Software as it was different in the Schematic
fix: Optimized Code in TX-Robot -> Removed some unused stuff
feat: Changed Wrinkle Detection Referencing system, as we had some troubles with the antistatic brush. -> Every TX-machine, which gets this version, need to recalibrate the wrinkle-Detection